Add: add controller for managin user data for auth. users and move some functionality to it from UserController

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -35,21 +36,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::find($request->id);
-        $user->details->contact_name = $request->contact_name;
-        $user->details->phone = $request->phone;
-        $user->details->mobile_1 = $request->mobile_1;
-        $user->details->mobile_2 = $request->mobile_2;
-        $user->details->mobile_3 = $request->mobile_3;
-        $user->details->viber = $request->viber;
-        $user->details->whats_up = $request->whats_up;
-        $user->details->telegram = $request->telegram;
-        $user->details->skype = $request->skype;
-        $user->details->vkontakte = $request->vkontakte;
-        $user->details->web_site = $request->web_site;
-        $user->details->instagram = $request->instagram;
-        $user->details->save();
-        $user->save();
+
     }
 
     /**
@@ -96,7 +83,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        $user->delete();
+        if(Auth::user()->role == 'admin')
+        {
+            $user = User::find($id);
+            $user->delete();
+        }
     }
 }
